working linter and prettier for sol and ts

+++ crypto/stark-verifier-ethereum/contracts/testing_contracts/public_coin_tester.sol
@@ -1,17 +1,16 @@
 pragma solidity ^0.6.4;
 
 import '../public_coin.sol';
-import '@nomiclabs/buidler/console.sol';
+
 
 contract PublicCoinTesting is PublicCoin {
-
     event log_bytes32(bytes32 data);
 
     // Reads a number of elements after initing to a starting value
-    function init_and_read(bytes calldata starting_data, uint number) external {
+    function init_and_read(bytes calldata starting_data, uint256 number) external {
         Coin memory coin = init_coin(starting_data);
 
-        for (uint i = 0; i < number; i++) {
+        for (uint256 i = 0; i < number; i++) {
             bytes32 read = read_bytes32(coin);
             emit log_bytes32(read);
         }
